# packages/frontend/backend_openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/courses": {
      "get": {
        "operationId": "CoursesController_getAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns all courses"
          }
        },
        "summary": "Get all courses",
        "tags": [
          "courses"
        ]
      },
      "post": {
        "operationId": "CoursesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Course created"
          }
        },
        "summary": "Create new course",
        "tags": [
          "courses"
        ]
      }
    },
    "/courses/{id}": {
      "get": {
        "operationId": "CoursesController_getOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Course ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns course"
          }
        },
        "summary": "Get course by ID",
        "tags": [
          "courses"
        ]
      },
      "patch": {
        "operationId": "CoursesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Course ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCourseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Course updated"
          }
        },
        "summary": "Update course",
        "tags": [
          "courses"
        ]
      },
      "delete": {
        "operationId": "CoursesController_deleteCourse",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Course ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Course deleted"
          }
        },
        "summary": "Delete course",
        "tags": [
          "courses"
        ]
      }
    },
    "/courses/batch-upsert": {
      "post": {
        "operationId": "CoursesController_batchUpsert",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchUpsertCourseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Courses created/updated. Returns count and the processed courses."
          }
        },
        "summary": "Batch create or update courses by name",
        "tags": [
          "courses"
        ]
      }
    },
...
    "/classes": {
      "get": {
        "operationId": "ClassesController_getAll",
        "parameters": [
          {
            "name": "groupId",
            "required": false,
            "in": "query",
            "description": "Filter by group ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns all classes or filtered by group ID"
          }
        },
        "summary": "Get all classes",
        "tags": [
          "classes"
        ]
      },
      "post": {
        "operationId": "ClassesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClassDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Class created"
          }
        },
        "summary": "Create new class",
        "tags": [
          "classes"
        ]
      }
    },
    "/classes/{id}": {
      "get": {
        "operationId": "ClassesController_getOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Class ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns class"
          }
        },
        "summary": "Get class by ID",
        "tags": [
          "classes"
        ]
      },
      "patch": {
        "operationId": "ClassesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Class ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClassDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Class updated"
          }
        },
        "summary": "Update class",
        "tags": [
          "classes"
        ]
      },
      "delete": {
        "operationId": "ClassesController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Class ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Class deleted"
          }
        },
        "summary": "Delete class",
        "tags": [
          "classes"
        ]
      }
    },
    "/classes/{id}/attendance": {
      "post": {
        "operationId": "ClassesController_setAttendance",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Class ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "number"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Attendance set, returns class and created debits"
          }
        },
        "summary": "Set attendance for a class",
        "tags": [
          "classes"
        ]
      }
    },
    "/classes/batch-upsert": {
      "post": {
        "operationId": "ClassesController_batchUpsert",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchUpsertClassDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Classes created/updated. Returns count and the processed classes."
          }
        },
        "summary": "Batch create or update classes by startTime and roomId",
        "tags": [
          "classes"
        ]
      }
    },
...
    "/auth/login": {
      "post": {
        "operationId": "PassportAuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "teacher@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "password123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful. Returns user data and sets access token cookie."
          },
          "401": {
            "description": "Unauthorized. Invalid credentials."
          }
        },
        "summary": "Login with email and password",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/auth/profile": {
      "get": {
        "operationId": "PassportAuthController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns current user profile."
          },
          "401": {
            "description": "Unauthorized. Valid JWT token required."
          }
        },
        "summary": "Get current user profile",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/auth/reset-password": {
      "post": {
        "operationId": "PassportAuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "teacher@example.com"
                  },
                  "newPassword": {
                    "type": "string",
                    "example": "newpassword123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset successful"
          },
          "404": {
            "description": "User not found"
          }
        },
        "summary": "Reset user password by email",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "KRUSANT API",
    "description": "API documentation for KRUSANT backend",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateCourseDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Course name",
            "example": "Course A"
          },
          "description": {
            "type": "string",
            "description": "Course description",
            "example": "Intensive course description"
          },
          "cost": {
            "type": "number",
            "description": "Course total cost",
            "example": 3325
          },
          "numberOfHours": {
            "type": "number",
            "description": "Total number of hours",
            "example": 70
          },
          "lessonLength": {
            "type": "string",
            "description": "Length of each lesson as time (HH:mm or HH:mm:ss)",
            "example": "02:30"
          },
          "pattern": {
            "type": "string",
            "description": "Pattern of classes",
            "example": "workdays",
            "enum": [
              "workdays",
              "weekends",
              "everyday",
              "weekly",
              "biweekly",
              "monthly"
            ]
          }
        },
        "required": [
          "name",
          "cost",
          "numberOfHours",
          "lessonLength",
          "pattern"
        ]
      },
      "UpdateCourseDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Course name",
            "example": "Course A"
          },
          "description": {
            "type": "string",
            "description": "Course description",
            "example": "Intensive course description"
          },
          "cost": {
            "type": "number",
            "description": "Course total cost",
            "example": 3325
          },
          "numberOfHours": {
            "type": "number",
            "description": "Total number of hours",
            "example": 70
          },
          "lessonLength": {
            "type": "string",
            "description": "Length of each lesson as time (HH:mm or HH:mm:ss)",
            "example": "02:30"
          },
          "pattern": {
            "type": "string",
            "description": "Pattern of classes",
            "example": "workdays",
            "enum": [
              "workdays",
              "weekends",
              "everyday",
              "weekly",
              "biweekly",
              "monthly"
            ]
          }
        }
      },
      "BatchUpsertCourseDto": {
        "type": "object",
        "properties": {
          "courses": {
            "description": "Array of courses to create or update",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateCourseDto"
            }
          }
        },
        "required": [
          "courses"
        ]
      },
...
      "CreateClassDto": {
        "type": "object",
        "properties": {
          "startTime": {
            "type": "string",
            "description": "Class start time (ISO datetime)",
            "example": "2026-01-06T10:00:00.000Z"
          },
          "lessonLength": {
            "type": "string",
            "description": "Lesson length as time (HH:mm or HH:mm:ss)",
            "example": "01:00"
          },
          "roomId": {
            "type": "number",
            "description": "Room ID",
            "example": 201
          },
          "groupId": {
            "type": "number",
            "description": "Group ID",
            "example": 10
          },
          "plannedStudentsIds": {
            "description": "IDs of planned students",
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "attendedStudentsIds": {
            "description": "IDs of attended students",
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "teacherId": {
            "type": "number",
            "description": "Teacher ID"
          },
          "cost": {
            "type": "number",
            "description": "Class cost",
            "example": 120
          },
          "comment": {
            "type": "string",
            "description": "Optional comment"
          }
        },
        "required": [
          "startTime",
          "lessonLength",
          "cost"
        ]
      },
      "UpdateClassDto": {
        "type": "object",
        "properties": {
          "startTime": {
            "type": "string",
            "description": "Class start time (ISO datetime)",
            "example": "2026-01-06T10:00:00.000Z"
          },
          "lessonLength": {
            "type": "string",
            "description": "Lesson length as time (HH:mm or HH:mm:ss)",
            "example": "01:00"
          },
          "roomId": {
            "type": "number",
            "description": "Room ID",
            "example": 201
          },
          "groupId": {
            "type": "number",
            "description": "Group ID",
            "example": 10
          },
          "plannedStudentsIds": {
            "description": "IDs of planned students",
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "attendedStudentsIds": {
            "description": "IDs of attended students",
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "teacherId": {
            "type": "number",
            "description": "Teacher ID"
          },
          "cost": {
            "type": "number",
            "description": "Class cost",
            "example": 120
          },
          "comment": {
            "type": "string",
            "description": "Optional comment"
          }
        }
      },
      "BatchUpsertClassDto": {
        "type": "object",
        "properties": {
          "classes": {
            "description": "Array of classes to create or update",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateClassDto"
            }
          }
        },
        "required": [
          "classes"
        ]
      },
...
    }
  }
}